refactor: separate language select
