+++ src/components/Header/LanguageSelect/LanguageSelect.scss
@@ -1,0 +1,5 @@
+@import "../../../styles/index.scss";
+
+.LanguagSelect {
+    display: flex;
+}
